+++ lib/ui/page/home/page/chat/widget/chat_forward.dart
@@ -43,6 +43,7 @@ import '/themes.dart';
 import '/ui/page/call/widget/fit_view.dart';
 import '/ui/page/home/page/chat/controller.dart';
 import '/ui/page/home/page/chat/forward/view.dart';
+import '/ui/page/home/page/user/controller.dart';
 import '/ui/page/home/widget/avatar.dart';
 import '/ui/widget/checkbox_button.dart';
 import '/ui/widget/context_menu/menu.dart';
@@ -109,7 +110,7 @@ class ChatForwardWidget extends StatefulWidget {
   /// [User] posted these [forwards].
   final RxUser? user;
 
-  /// Indicator whether this [ChatForwardWidget] should display [RxUser.title].
+  /// Indicator whether this [ChatForwardWidget] should display [UserExt.title].
   ///
   /// For example, [Chat]-groups should display messages with titles.
   final bool withName;
@@ -341,7 +342,7 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                           padding: const EdgeInsets.fromLTRB(12, 0, 9, 0),
                           child: SelectionText.rich(
                             TextSpan(
-                              text: widget.user?.title ?? 'dot'.l10n * 3,
+                              text: widget.user?.title() ?? 'dot'.l10n * 3,
                               recognizer: TapGestureRecognizer()
                                 ..onTap = () =>
                                     router.user(widget.authorId, push: true),
@@ -469,7 +470,7 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                       child: Padding(
                         padding: const EdgeInsets.only(right: 9),
                         child: SelectionText(
-                          user?.title ?? 'dot'.l10n * 3,
+                          user?.title() ?? 'dot'.l10n * 3,
                           selectable:
                               widget.selectable &&
                               (PlatformUtils.isDesktop || menu),
@@ -744,7 +745,7 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
                     padding: const EdgeInsets.fromLTRB(12, 0, 9, 0),
                     child: SelectionText.rich(
                       TextSpan(
-                        text: widget.user?.title ?? 'dot'.l10n * 3,
+                        text: widget.user?.title() ?? 'dot'.l10n * 3,
                         recognizer: TapGestureRecognizer()
                           ..onTap = () =>
                               router.user(widget.authorId, push: true),
@@ -877,7 +878,7 @@ class _ChatForwardWidgetState extends State<ChatForwardWidget> {
               futureOr: () => widget.getUser?.call(m.memberId),
               builder: (context, member) {
                 return Tooltip(
-                  message: member?.title ?? user?.title ?? ('dot'.l10n * 3),
+                  message: member?.title() ?? user?.title() ?? ('dot'.l10n * 3),
                   verticalOffset: 15,
                   padding: const EdgeInsets.fromLTRB(7, 3, 7, 3),
                   decoration: BoxDecoration(