+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All user visible changes to this project will be documented in this file. This p
 ### Changed
 
 - UI:
+    - Home page:
+        - Display "Deleted Account" title for deleted users. ([#1445], [#1419])
     - Chats tab:
         - Redesigned app bar and searching. ([#1438], [#1396])
     - Chat page:
@@ -27,8 +29,11 @@ All user visible changes to this project will be documented in this file. This p
         - Invalid tooltip positions for buttons in dock. ([#1473], [#1436])
 
 [#1396]: /../../issues/1396
+[#1419]: /../../issues/1419
+[#1436]: /../../issues/1436
 [#1436]: /../../issues/1436
 [#1438]: /../../pull/1438
+[#1445]: /../../pull/1445
 [#1453]: /../../issues/1453
 [#1466]: /../../pull/1466
 [#1473]: /../../pull/1473